requirements.txt
OpenOPC-Python3x
pypiwin32
clickhouse_driver
requests
pandas
matplotlib